synchronize read me and add demo google account
