+++ README.md
@@ -20,17 +20,17 @@
 
 ##### Description
 
-A travel planning website that ease the hassel of planning travel schedule by combining map direction, calendar timeline and travel blog with a beautiful and easy-to-use drag and drop interface.
+A travel planning website that eases the hassle of planning travel schedules by combining map direction, calendar timeline and travel blog with a beautiful and easy-to-use drag and drop interface.
 
 Enables users to share, collect and import travel schedules made by others.
 
 <details> 
-<summary><em> Reason behind this project </em></summary> 
+<summary><em> The reason behind this project </em></summary> 
 <blockquote>
-<p> Pandemic has stopped a lot of travel plans in the past few years. Recently locations around the world are beginning to open for travels, I think it's the perfect timing to build a travel planning website.  :wink: </p>
+<p> Pandemic has stopped a lot of travel plans in the past few years. Recently locations around the world are beginning to open for travel, I think it's the perfect timing to build a travel planning website.  :wink: </p>
 
 <p>
-I've been planning my trips with spreadsheets and note takers, in which I had to switch platforms between calendars, maps and a lot times many personal travel blogs for planning ideas.
+I've been planning my trips with spreadsheets and note-takers, in which I had to switch platforms between calendars, maps and a lot of times many personal travel blogs for planning ideas.
 
 So here it is, **TravelBlox** that puts everything altogether :blush: :sparkling_heart:
 
@@ -45,12 +45,23 @@ So here it is, **TravelBlox** that puts everything altogether :blush: :sparkling
 > Link: https://travelblox-a8094.web.app/
 
 - Demo Account:
+
   - Username: user@user.com
   - Password: user1234
 
+- Demo Account for Google Calendar:
+  - Username:demopurposeuser@gmail.com
+  - Password: demouser1234
+
+<em>
+*This application is still in review by Google for release, which takes 4-6 weeks, so you may not be able to export it to your google calendar yet.
+<br>
+*For demo purposes, please use the provided Google account for Google Calendar 
+</em>
+
 <br />
 
-## Technologies
+## Technologies Used
 
 #### Front-End
 
@@ -75,7 +86,7 @@ So here it is, **TravelBlox** that puts everything altogether :blush: :sparkling
 
 - SVG / Adobe Illustrator / Adobe Photoshop
 
-#### Third Party APIs
+#### Third-Party APIs
 
 - Google's Maps API, Directions API, Places API, Autocomplete API, Calendar API
 - OpenWeather API, REST Countries API, Travel Advisor API
@@ -84,15 +95,53 @@ So here it is, **TravelBlox** that puts everything altogether :blush: :sparkling
 
 ## Flow Chart
 
-<img align="center"  alt="TravelBlox Logo" src="./read_me/Userflow.png" >
+<img align="center"  alt="Userflow Image" src="./read_me/userflow-image.png" >
+<br />
+
+### User journey and features
+
+<br />
+
+#### User sign up / sign in
+
+- Implemented parallax and animation effects
+- Authorized native, Google and Facebook Login
+- Discover travel plans shared by others
+
+ <img align="center"  alt="TravelBlox Logo" src="./read_me/login-optimized.gif" >
+
+#### Favourite a travel plan
+
+- Able to search through discoverable plans by author name, plan title or country
+- Furnished schedule events' information details with **calendar view** in day, week, month or agenda, along with **map directions** and a **blog** section for user-uploaded images and descriptions
+- Favourite travel plans that could be later <em>imported and customized</em> into your own travel schedule
+
+ <img align="center"  alt="TravelBlox Logo" src="./read_me/discover-and-favourite-optimized.gif" >
+
+#### Import travel schedules and Search through attractions
+
+- **Import** events from plans that you've favourited and add your own images or description
+- Facilitated schedule arrangement by **draggable** event time blocks
+- Get ideas from Tripadvisor's ranking on **attractions and restaurants** of that city
+
+<img align="center"  alt="TravelBlox Logo" src="./read_me/import-search-attraction-optimized.gif" >
+
+- And there you have it! Your own customized travel schedule
+  <img align="center"  alt="TravelBlox Logo" src="./read_me/check-static-after-import-optimized.gif" >
+
+#### Export to Google Calendar
+
+- Integrate with your timetable by connecting and exporting to your Google Calendar
+  <img align="center"  alt="TravelBlox Logo" src="./read_me/export-to-calendar.gif" >
+
 <br />
 
 ## Future Features
 
-- Follow and Add friend system
+- Follow and Add friends system
 - Co-editing with friends
-- Commenting system
-- Export a plan as printable PDF booklet
+- Commenting on plans
+- Export schedules as printable PDF booklets
 
 <br />
