+++ src/utils/googleServices.js
@@ -1,7 +1,6 @@
 export const googleServices = {
   getlaceDetail(placeId) {
     let map = new window.google.maps.Map(document.createElement('div'));
-    console.log(22, map);
     return new Promise(function (resolve, reject) {
       let placesService = new window.google.maps.places.PlacesService(map);
       placesService.getDetails(
@@ -11,7 +10,6 @@ export const googleServices = {
         },
         (place) => {
           resolve(place);
-          console.log(33, place);
           return place;
         }
       );
